requirements.txt
pyinstaller
kivy
opencv-python
pycryptodome
#onnxruntime # uninstalled 
matplotlib
scikit-image
filterpy
tensorflow